chore(contracts): move inline tests to a separate file
The test module was extracted from contracts.rs into its own file, contracts_tests.rs, to reduce the size of the main source file and improve readability. The module declaration now uses the `#[path]` attribute to reference the external test file.

Auto-committed-on: dragonfly
